+++ test/test_friend_collection_in_batches.py
@@ -1,6 +1,6 @@
 
 
-from app.workers.friend_batch_collector import split_into_batches
+from app.app.friend_graphs.collect_in_batches import split_into_batches
 
 def test_split_into_batches():
     batches = split_into_batches([0,1,2,3,4,5,6,7,8,9,10], 3)
